feat: add endpoints for creating data resources
when a user POSTs to /analytics, a task is created, responsible for creating a file that holds the requested data. this task and its status can be queried. a middleware pushes this file to the space it was requested from. commit also reconfigures code in line with more standard express app structures

closes #13